fix: update import paths to include extension
The import path must include the extension when publishing ESM as
otherwise the module loader doesn't know how to find the file.

Fixes errors like:

```
Relative import paths need explicit file extensions in EcmaScript imports when '--moduleResolution' is 'node16' or 'nodenext'. Consider adding an
extension to the import path.
````

This is a weakness of JSDoc types and would be caught by converting
to TypeScript.

Refs: https://github.com/multiformats/js-multiformats/issues/249